backend/requirements.txt
# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis & Celery
redis==5.0.1
celery==5.3.6
httpx==0.26.0

# Neo4j
neo4j==5.18.0

# LangChain & LangGraph
langchain==0.1.17
langchain-community==0.0.29
langgraph==0.0.35
langchain-core==0.1.52

# LLM & Embeddings
openai==1.12.0
tiktoken==0.5.2

# Document Processing
python-docx==1.1.0
python-pptx==0.6.23
pypdf==4.0.1
markdown==3.5.2

# MinIO / S3
minio==7.2.3
urllib3==1.26.18

# Utils
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyYAML==6.0.1
tenacity==8.2.3
aiofiles==23.2.1

# Supabase
supabase==2.3.7
